Rename (content-blocks) : import path 수정
