+++ app/components/contents/Callout.tsx
@@ -1,7 +1,7 @@
 import { CalloutBlockObjectResponse } from "@notionhq/client/build/src/api-endpoints";
 
 import { fetchBlocks } from "@/app/lib/databases";
-import RichText from "./RichText";
+import RichText from "./texts/RichText";
 import Block from "./Block";
 
 export default async function Callout({ block }: { block: CalloutBlockObjectResponse }) {
