+++ app/components/contents/Equation.tsx
@@ -1,5 +1,6 @@
-import { renderKatex } from "@/app/lib/functions/katex";
 import { EquationBlockObjectResponse } from "@notionhq/client/build/src/api-endpoints";
+
+import { renderKatex } from "@/app/lib/functions/katex";
 
 export default function Equation({ block }: { block: EquationBlockObjectResponse }) {
     const katexHtml = renderKatex(block.equation.expression);
