+++ app/components/contents/tables/TableData.tsx
@@ -1,6 +1,6 @@
 import { RichTextItemResponse } from "@notionhq/client/build/src/api-endpoints";
 
-import RichText from "../RichText";
+import RichText from "../texts/RichText";
 
 export default function TableData({ cell }: { cell: RichTextItemResponse[] }) {
     return (
