+++ app/components/contents/texts/RichText.tsx
@@ -2,8 +2,8 @@ import { RichTextItemResponse } from "@notionhq/client/build/src/api-endpoints";
 
 import { extractRichTextStyle } from "@/app/lib/functions/notion";
 import { renderKatex } from "@/app/lib/functions/katex";
-import CodeText from "./texts/CodeText";
-import LinkText from "./texts/LinkText";
+import CodeText from "./CodeText";
+import LinkText from "./LinkText";
 
 export default function RichText({ richTexts }: { richTexts: RichTextItemResponse[] }) {
     return richTexts.map((richText, index) => {
